+++ content/pages/collections.mdx
@@ -2,4 +2,13 @@
 title: Collections
 ---
 
-Check out my dope ass shoes.
+## Shoes
+
+- Nike Blazer Mid '77 Vintage 'Have A Good Game'
+- Adidas LEGO x ZX 8000 'A-ZX Series - Mixed Blocks'
+- Puma Nintendo x J. Cole x RS-Dreamer 'Super Mario 64'
+- Adidas Marvel x D.O.N. Issue #2 GCA 'Spida Sense'
+- Reebok U.S.C.M. Bug x Alien Stomper '40th Anniversary'
+- Nike Adapt Huarache 'Opti Yellow'
+- Vans Disney x Sk8-Hi 'Sorcerer's Apprentice'
+- Adidas Marvel x Speedfactory AM4CM 'Captain Marvel'
